pyproject.toml
[tool.poetry]
name = "ocm-sandbox"
version = "0.1.0"
description = "Local development tools for Open Cluster Management"
authors = ["Scott Brimhall"]
readme = "README.md"
homepage = "https://github.com/suchpuppet/ocm-kind-sandbox"
repository = "https://github.com/suchpuppet/ocm-kind-sandbox"
keywords = ["kubernetes", "ocm", "open-cluster-management", "kind", "multicluster"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "ocm_sandbox"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
pyyaml = "^6.0"
rich = "^13.7.0"  # For beautiful terminal output (used by typer)

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
pylint = "^3.0.0"
flake8 = "^6.1.0"
black = "^23.12.0"  # Code formatter
isort = "^5.13.0"  # Import sorter

[tool.poetry.scripts]
ocm-sandbox = "ocm_sandbox.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true

[tool.pylint.messages_control]
max-line-length = 120
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0913",  # too-many-arguments
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=ocm_sandbox --cov-report=term --cov-report=html"
